python/pyproject.toml
[project]
name = "cross-docs"
version = "0.1.0"
description = "Documentation framework built on Cross-Inertia"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Patrick Arminio", email = "patrick.arminio@gmail.com" }]
keywords = ["documentation", "fastapi", "inertia", "react", "ssr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Documentation",
    "Topic :: Internet :: WWW/HTTP",
]

dependencies = [
    "fastapi>=0.100.0",
    "cross-inertia>=0.10.0",
]

[project.urls]
Homepage = "https://github.com/usecross/cross-docs"
Documentation = "https://cross-docs.dev"
Repository = "https://github.com/usecross/cross-docs"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cross_docs"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
